Refactor: Moving `profileViewBinding` and `profileViewModel` from `MainActivity` to `ProfilePageFragment`. Clear main activity.
